# docs/swagger.yaml
basePath: /api
definitions:
  backend_go_internal_model.DeckType:
    enum:
    - fibonacci
    - modified_fibonacci
    - tshirt
    - hydra
    - classic
    type: string
    x-enum-varnames:
    - DeckTypeFibonacci
    - DeckTypeModifiedFibonacci
    - DeckTypeTShirt
    - DeckTypeHydra
    - DeckTypeClassic
  backend_go_internal_model_apimodel.CreateSessionResponse:
    description: Данные созданной сессии
    properties:
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      session_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
    type: object
  backend_go_internal_model_apimodel.DeleteSessionResponse:
    description: Результат операции удаления сессии
    properties:
      message:
        example: Сессия успешно удалена
        type: string
      success:
        example: true
        type: boolean
    type: object
  backend_go_internal_model_apimodel.GuestLogin:
    description: Данные для создания гостевого аккаунта
    properties:
      name:
        example: Guest User
        type: string
    type: object
  backend_go_internal_model_apimodel.SessionCreate:
    properties:
      deck_type:
        $ref: '#/definitions/backend_go_internal_model.DeckType'
      name:
        type: string
    type: object
  backend_go_internal_model_apimodel.TokenResponse:
    description: Токены доступа и обновления
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      token_type:
        example: Bearer
        type: string
    type: object
  backend_go_internal_model_apimodel.UserLogin:
    description: Данные для входа пользователя
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: securepassword123
        type: string
    type: object
  backend_go_internal_model_apimodel.UserProfile:
    description: Полная информация о пользователе
    properties:
      avatar_url:
        example: https://example.com/avatar.jpg
        type: string
      email:
        example: john@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      is_active:
        example: true
        type: boolean
      is_guest:
        example: false
        type: boolean
      is_verified:
        example: false
        type: boolean
      name:
        example: John Doe
        type: string
      oauth_provider:
        example: google
        type: string
    type: object
  backend_go_internal_model_apimodel.UserRegister:
    description: Данные для регистрации нового пользователя
    properties:
      email:
        example: john@example.com
        type: string
      name:
        example: John Doe
        type: string
      password:
        example: securepassword123
        type: string
    type: object
  backend_go_internal_model_entitymodel.Session:
    properties:
      allow_emoji:
        type: boolean
      auto_reveal:
        type: boolean
      cards_revealed:
        type: boolean
      created_at:
        type: string
      created_via:
        type: string
      creator_id:
        type: string
      creator_name:
        type: string
      deck_type:
        $ref: '#/definitions/backend_go_internal_model.DeckType'
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  internal_api_handler.ErrorResponse:
    description: Сообщение об ошибке
    properties:
      error:
        example: invalid credentials
        type: string
    type: object
info:
  contact:
    email: kda94@mail.ru
    name: Dmitriy Kazakov
  description: API для управления сессиями планирования с системой аутентификации
    и WebSocket
  title: Backend Go API
  version: "1.0"
paths:
  /auth/guest_login:
    post:
      consumes:
      - application/json
      description: Создание временной гостевой учетной записи
      parameters:
      - description: Данные гостя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/backend_go_internal_model_apimodel.GuestLogin'
      produces:
      - application/json
      responses:
        "200":
          description: Успешная аутентификация гостя
          schema:
            $ref: '#/definitions/backend_go_internal_model_apimodel.TokenResponse'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      summary: Позволяет войти как гость
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Аутентификация пользователя по email и паролю
      parameters:
      - description: Данные для входа
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/backend_go_internal_model_apimodel.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: Успешная аутентификация
          schema:
            $ref: '#/definitions/backend_go_internal_model_apimodel.TokenResponse'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "401":
          description: Неверные учетные данные
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      summary: Вход в систему
      tags:
      - auth
  /auth/me:
    get:
      consumes:
      - application/json
      description: Получение информации об авторизованном пользователе
      produces:
      - application/json
      responses:
        "200":
          description: Данные пользователя
          schema:
            $ref: '#/definitions/backend_go_internal_model_apimodel.UserProfile'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить текущего пользователя
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Создание нового аккаунта пользователя
      parameters:
      - description: Данные для регистрации
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/backend_go_internal_model_apimodel.UserRegister'
      produces:
      - application/json
      responses:
        "200":
          description: Успешная регистрация
          schema:
            $ref: '#/definitions/backend_go_internal_model_apimodel.TokenResponse'
        "400":
          description: Пользователь уже существует или ошибка валидации
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      summary: Регистрация пользователя
      tags:
      - auth
  /sessions:
    get:
      consumes:
      - application/json
      description: Получение списка сессий, созданных текущим пользователем
      produces:
      - application/json
      responses:
        "200":
          description: Список сессий пользователя
          schema:
            items:
              $ref: '#/definitions/backend_go_internal_model_entitymodel.Session'
            type: array
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "403":
          description: Гостевой пользователь не имеет доступа
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить мои сессии
      tags:
      - sessions
    post:
      consumes:
      - application/json
      description: Создание новой сессии для планирования с возможностью голосования
      parameters:
      - description: Данные для создания сессии
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/backend_go_internal_model_apimodel.SessionCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Сессия успешно создана
          schema:
            $ref: '#/definitions/backend_go_internal_model_apimodel.CreateSessionResponse'
        "400":
          description: Ошибка валидации данных
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать сессию
      tags:
      - sessions
  /sessions/{session_id}:
    delete:
      consumes:
      - application/json
      description: Удаление сессии по ID (только для создателя сессии)
      parameters:
      - description: UUID сессии
        in: path
        name: session_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Сессия успешно удалена
          schema:
            $ref: '#/definitions/backend_go_internal_model_apimodel.DeleteSessionResponse'
        "401":
          description: Не авторизован
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "403":
          description: Доступ запрещен (гостевой пользователь или не создатель сессии)
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "404":
          description: Сессия не найдена
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить сессию
      tags:
      - sessions
    get:
      consumes:
      - application/json
      description: Получение информации о конкретной сессии по её идентификатору
      parameters:
      - description: UUID сессии
        in: path
        name: session_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Данные сессии
          schema:
            $ref: '#/definitions/backend_go_internal_model_entitymodel.Session'
        "404":
          description: Сессия не найдена
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/internal_api_handler.ErrorResponse'
      summary: Получить сессию по ID
      tags:
      - sessions
securityDefinitions:
  BearerAuth:
    description: 'JWT токен в формате: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
tags:
- description: Аутентификация и управление пользователями
  name: auth
- description: Управление сессиями планирования и голосования
  name: sessions